Split out string object and String veneer class into strng.c/.h
