+++ src/builtin/veneer.c
@@ -261,194 +261,6 @@ MORPHO_METHOD(MORPHO_CLONE_METHOD, Object_clone, BUILTIN_FLAGSEMPTY)
 MORPHO_ENDCLASS
 
 /* **********************************************************************
- * String
- * ********************************************************************** */
-
-/** Convert a string to a number */
-bool string_tonumber(objectstring *string, value *out) {
-    bool minus=false;
-    lexer l;
-    token tok;
-    error err;
-    lex_init(&l, string->string, 0);
-
-    if (lex(&l, &tok, &err)) {
-        if (tok.type==TOKEN_MINUS) { // Check for leading minus
-            minus=true;
-            if (!lex(&l, &tok, &err)) return false;
-        } else if (tok.type==TOKEN_PLUS) { // or plus
-            if (!lex(&l, &tok, &err)) return false;
-        }
-
-        if (tok.type==TOKEN_INTEGER) {
-            long i = strtol(tok.start, NULL, 10);
-            if (minus) i=-i;
-            *out = MORPHO_INTEGER((int) i);
-            return true;
-        } else if (tok.type==TOKEN_NUMBER) {
-            double f = strtod(tok.start, NULL);
-            if (minus) f=-f;
-            *out = MORPHO_FLOAT(f);
-            return true;
-        }
-    }
-
-    return false;
-}
-
-/** Count number of characters in a string */
-int string_countchars(objectstring *s) {
-    int n=0;
-    for (uint8_t *c = (uint8_t *) s->string; *c!='\0'; ) {
-        c+=morpho_utf8numberofbytes(c);
-        n++;
-    }
-    return n;
-}
-
-/** Get a pointer to the i'th character of a string */
-char *string_index(objectstring *s, int i) {
-    int n=0;
-    for (uint8_t *c = (uint8_t *) s->string; *c!='\0'; ) {
-        if (i==n) return (char *) c;
-        c+=morpho_utf8numberofbytes(c);
-        n++;
-    }
-    return NULL;
-}
-
-/** Constructor */
-value string_constructor(vm *v, int nargs, value *args) {
-    value out=morpho_concatenate(v, nargs, args+1);
-    if (MORPHO_ISOBJECT(out)) morpho_bindobjects(v, 1, &out);
-    return out;
-}
-
-/** Find a string's length */
-value String_count(vm *v, int nargs, value *args) {
-    objectstring *slf = MORPHO_GETSTRING(MORPHO_SELF(args));
-
-    return MORPHO_INTEGER(string_countchars(slf));
-}
-
-/** Prints a string */
-value String_print(vm *v, int nargs, value *args) {
-    morpho_printvalue(MORPHO_SELF(args));
-
-    return MORPHO_SELF(args);
-}
-
-/** Clones a string */
-value String_clone(vm *v, int nargs, value *args) {
-    objectstring *slf = MORPHO_GETSTRING(MORPHO_SELF(args));
-    value out = object_stringfromcstring(slf->string, slf->length);
-    if (MORPHO_ISNIL(out)) morpho_runtimeerror(v, ERROR_ALLOCATIONFAILED);
-    morpho_bindobjects(v, 1, &out);
-    return out;
-}
-
-/** Sets an index */
-/*value String_setindex(vm *v, int nargs, value *args) {
-    objectstring *slf = MORPHO_GETSTRING(MORPHO_SELF(args));
-
-    morpho_runtimeerror(v, STRING_IMMTBL);
-
-    if (nargs==2 &&
-        MORPHO_ISINTEGER(MORPHO_GETARG(args, 0)) &&
-        MORPHO_ISSTRING(MORPHO_GETARG(args, 1))) {
-        int n=MORPHO_GETINTEGERVALUE(MORPHO_GETARG(args, 0));
-        objectstring *set = MORPHO_GETSTRING(MORPHO_GETARG(args, 1));
-
-        if (n>=0 && n<slf->length) {
-            for (unsigned int i=0; i<set->length && n+i<slf->length; i++) {
-                slf->stringdata[n+i]=set->stringdata[i];
-            }
-        } else morpho_runtimeerror(v, VM_OUTOFBOUNDS);
-    } else morpho_runtimeerror(v, SETINDEX_ARGS);
-
-    return MORPHO_NIL;
-}*/
-
-/** Enumerate members of a string */
-value String_enumerate(vm *v, int nargs, value *args) {
-    objectstring *slf = MORPHO_GETSTRING(MORPHO_SELF(args));
-    value out=MORPHO_NIL;
-
-    if (nargs==1 && MORPHO_ISINTEGER(MORPHO_GETARG(args, 0))) {
-        int n=MORPHO_GETINTEGERVALUE(MORPHO_GETARG(args, 0));
-
-        if (n<0) {
-            out=MORPHO_INTEGER(string_countchars(slf));
-        } else {
-            char *c=string_index(slf, n);
-            if (c) {
-                out=object_stringfromcstring(c, morpho_utf8numberofbytes((uint8_t *) c));
-                morpho_bindobjects(v, 1, &out);
-            } else morpho_runtimeerror(v, VM_OUTOFBOUNDS);
-        }
-    } else MORPHO_RAISE(v, ENUMERATE_ARGS);
-
-    return out;
-}
-
-value String_isnumber(vm *v, int nargs, value *args) {
-    objectstring *slf = MORPHO_GETSTRING(MORPHO_SELF(args));
-    value out=MORPHO_NIL;
-
-    if (string_tonumber(slf, &out)) return MORPHO_TRUE;
-
-    return MORPHO_FALSE;
-}
-
-value String_split(vm *v, int nargs, value *args) {
-    objectstring *slf = MORPHO_GETSTRING(MORPHO_SELF(args));
-    value out=MORPHO_NIL;
-
-    if (nargs==1 && MORPHO_ISSTRING(MORPHO_GETARG(args, 0))) {
-        objectstring *split = MORPHO_GETSTRING(MORPHO_GETARG(args, 0));
-        objectlist *new = object_newlist(0, NULL);
-
-        if (!new) { morpho_runtimeerror(v, ERROR_ALLOCATIONFAILED); return MORPHO_NIL; }
-
-        char *last = slf->string;
-        for (char *c = slf->string; *c!='\0'; c+=morpho_utf8numberofbytes((uint8_t *) c)) { // Loop over string
-            for (char *s = split->string; *s!='\0';) { // Loop over split chars
-                int nbytes = morpho_utf8numberofbytes((uint8_t *) s);
-                if (strncmp(c, s, nbytes)==0) {
-                    value newstring = object_stringfromcstring(last, c-last);
-                    if (MORPHO_ISNIL(newstring)) morpho_runtimeerror(v, ERROR_ALLOCATIONFAILED);
-                    list_append(new, newstring);
-                    last=c+nbytes;
-                }
-                s+=nbytes;
-            }
-        }
-
-        value newstring = object_stringfromcstring(last, slf->string+slf->length-last);
-        if (MORPHO_ISNIL(newstring)) morpho_runtimeerror(v, ERROR_ALLOCATIONFAILED);
-        list_append(new, newstring);
-
-        out=MORPHO_OBJECT(new);
-        list_append(new, out);
-        morpho_bindobjects(v, new->val.count, new->val.data);
-        new->val.count-=1;
-    }
-
-    return out;
-}
-
-MORPHO_BEGINCLASS(String)
-MORPHO_METHOD(MORPHO_COUNT_METHOD, String_count, BUILTIN_FLAGSEMPTY),
-MORPHO_METHOD(MORPHO_PRINT_METHOD, String_print, BUILTIN_FLAGSEMPTY),
-MORPHO_METHOD(MORPHO_CLONE_METHOD, String_clone, BUILTIN_FLAGSEMPTY),
-MORPHO_METHOD(MORPHO_GETINDEX_METHOD, String_enumerate, BUILTIN_FLAGSEMPTY),
-//MORPHO_METHOD(MORPHO_SETINDEX_METHOD, String_setindex, BUILTIN_FLAGSEMPTY),
-MORPHO_METHOD(MORPHO_ENUMERATE_METHOD, String_enumerate, BUILTIN_FLAGSEMPTY),
-MORPHO_METHOD(STRING_ISNUMBER_METHOD, String_isnumber, BUILTIN_FLAGSEMPTY),
-MORPHO_METHOD(STRING_SPLIT_METHOD, String_split, BUILTIN_FLAGSEMPTY)
-MORPHO_ENDCLASS
-
-/* **********************************************************************
  * Array
  * ********************************************************************** */
 
@@ -1361,11 +1173,6 @@ void veneer_initialize(void) {
     value objclass=builtin_addclass(OBJECT_CLASSNAME, MORPHO_GETCLASSDEFINITION(Object), MORPHO_NIL);
     morpho_setbaseclass(objclass);
 
-    /* String */
-    builtin_addfunction(STRING_CLASSNAME, string_constructor, BUILTIN_FLAGSEMPTY);
-    value stringclass=builtin_addclass(STRING_CLASSNAME, MORPHO_GETCLASSDEFINITION(String), objclass);
-    object_setveneerclass(OBJECT_STRING, stringclass);
-
     /* Array */
     builtin_addfunction(ARRAY_CLASSNAME, array_constructor, BUILTIN_FLAGSEMPTY);
     value arrayclass=builtin_addclass(ARRAY_CLASSNAME, MORPHO_GETCLASSDEFINITION(Array), objclass);
@@ -1397,7 +1204,6 @@ void veneer_initialize(void) {
     morpho_defineerror(ARRAY_ARGS, ERROR_HALT, ARRAY_ARGS_MSG);
     morpho_defineerror(ARRAY_INIT, ERROR_HALT, ARRAY_INIT_MSG);
     morpho_defineerror(ARRAY_CMPT, ERROR_HALT, ARRAY_CMPT_MSG);
-    morpho_defineerror(STRING_IMMTBL, ERROR_HALT, STRING_IMMTBL_MSG);
     morpho_defineerror(ENUMERATE_ARGS, ERROR_HALT, ENUMERATE_ARGS_MSG);
     morpho_defineerror(DICT_DCTKYNTFND, ERROR_HALT, DICT_DCTKYNTFND_MSG);
     morpho_defineerror(DICT_DCTSTARG, ERROR_HALT, DICT_DCTSTARG_MSG);
