+++ src/classes/CMakeLists.txt
@@ -4,5 +4,6 @@ target_sources(morpho
         file.c   file.h
         list.c   list.h
         range.c  range.h
+        strng.c  strng.h
         system.c system.h
 )
