+++ src/vm/profile.c
@@ -5,6 +5,7 @@
  */
 
 #include "profile.h"
+#include "strng.h"
 
 /* **********************************************************************
 * Profiler
